requirements.txt
absl-py==0.8.1
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
cloudpickle==1.1.1
contextlib2==21.6.0
coverage==6.4.2
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dm-sonnet==1.36
dm-tree==0.1.1
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.1
filelock==3.7.1
flake8==3.9.2
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-import-style==0.5.0
fonttools==4.34.4
gast==0.2.2
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
html2text==2020.1.16
identify==2.5.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonlines==3.0.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
m2r2==0.3.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nodeenv==1.7.0
notebook==6.4.12
numpy==1.18.1
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pockets==0.9.1
pre-commit==2.19.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pydash==5.1.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-assume==2.4.3
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
semantic-version==2.10.0
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.0
sphinx-markdown-builder==0.5.5
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-hub==0.12.0
tensorflow-probability==0.8.0
termcolor==1.1.0
terminado==0.15.0
tf-slim==1.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
unify==0.5
untokenize==0.1.1
urllib3==1.26.11
virtualenv==20.16.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wrapt==1.14.1
XlsxWriter==3.0.3
yapf==0.32.0
zipp==3.8.1
